+++ CHANGELOG.md
@@ -9,9 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `PayIdClient` provides the functionality of `PayIDClient` under an idiomatic cased name.
+- `PayIdUtils` provides the functionality of `PayIDUtils` under an idiomatic cased name.
 
 ### Deprecated
 - `PayIDClient` is deprecated, use the idiomatically named `PayIdClient` instead.
+- `PayIDUtils` is deprecated, use the idiomatically named `PayIdUtils` instead.
 
 ## 5.2.0 - 2020-06-01